remove 'legacy_renum_only' flag and move the function 'decompress_to_edgelist'
