+++ python/cugraph/cugraph/structure/__init__.py
@@ -30,7 +30,6 @@ from cugraph.structure.replicate_edgelist import (
     replicate_cudf_dataframe,
     replicate_cudf_series,
 )
-from cugraph.structure.decompress_to_edgelist import decompress_to_edgelist
 from cugraph.structure.convert_matrix import (
     from_edgelist,
     from_cudf_edgelist,
