+++ python/cugraph/cugraph/structure/graph_classes.py
@@ -115,7 +115,6 @@ class Graph:
         edge_type=None,
         renumber=True,
         store_transposed=False,
-        legacy_renum_only=False,
         symmetrize=None,
     ):
         """
@@ -168,13 +167,6 @@ class Graph:
             If True, stores the transpose of the adjacency matrix.  Required
             for certain algorithms.
 
-        legacy_renum_only : bool, optional (default=False)
-            If True, skips the C++ renumbering step.  Must be true for
-            pylibcugraph algorithms.  Must be false for algorithms
-            not yet converted to the pylibcugraph C API.
-
-            This parameter is deprecated and will be removed.
-
         symmetrize: bool, optional (default=None)
             If True, symmetrize the edge list for an undirected graph. Setting
             this flag to True for a directed graph returns an error. The default
@@ -210,7 +202,6 @@ class Graph:
             edge_type=edge_type,
             renumber=renumber,
             store_transposed=store_transposed,
-            legacy_renum_only=legacy_renum_only,
             symmetrize=symmetrize,
         )
 
@@ -305,8 +296,7 @@ class Graph:
         edge_id=None,
         edge_type=None,
         renumber=True,
-        store_transposed=False,
-        legacy_renum_only=False,
+        store_transposed=False
     ):
         """
         Initializes the distributed graph from the dask_cudf.DataFrame
@@ -353,13 +343,6 @@ class Graph:
             If True, stores the transpose of the adjacency matrix.  Required
             for certain algorithms.
 
-        legacy_renum_only : bool, optional (default=False)
-            If True, skips the C++ renumbering step.  Must be true for
-            pylibcugraph algorithms.  Must be false for algorithms
-            not yet converted to the pylibcugraph C API.
-
-            This parameter is deprecated and will be removed.
-
         """
 
         if self._Impl is None:
@@ -377,8 +360,7 @@ class Graph:
             edge_id=edge_id,
             edge_type=edge_type,
             renumber=renumber,
-            store_transposed=store_transposed,
-            legacy_renum_only=legacy_renum_only,
+            store_transposed=store_transposed
         )
 
     # Move to Compat Module
@@ -868,8 +850,7 @@ class NPartiteGraph(Graph):
         destination="destination",
         edge_attr=None,
         renumber=True,
-        store_transposed=False,
-        legacy_renum_only=False,
+        store_transposed=False
     ):
         """
         Initialize a graph from the edge list. It is an error to call this
@@ -909,13 +890,6 @@ class NPartiteGraph(Graph):
             If True, stores the transpose of the adjacency matrix.  Required
             for certain algorithms.
 
-        legacy_renum_only : bool, optional (default=False)
-            If True, skips the C++ renumbering step.  Must be true for
-            pylibcugraph algorithms.  Must be false for algorithms
-            not yet converted to the pylibcugraph C API.
-
-            This parameter is deprecated and will be removed.
-
         Examples
         --------
         >>> df = cudf.read_csv(datasets_path / 'karate.csv', delimiter=' ',
@@ -944,8 +918,7 @@ class NPartiteGraph(Graph):
         destination="destination",
         edge_attr=None,
         renumber=True,
-        store_transposed=False,
-        legacy_renum_only=False,
+        store_transposed=False
     ):
         """
         Initializes the distributed graph from the dask_cudf.DataFrame
@@ -980,12 +953,6 @@ class NPartiteGraph(Graph):
             If True, stores the transpose of the adjacency matrix.  Required
             for certain algorithms.
 
-        legacy_renum_only : bool, optional (default=False)
-            If True, skips the C++ renumbering step.  Must be true for
-            pylibcugraph algorithms.  Must be false for algorithms
-            not yet converted to the pylibcugraph C API.
-
-            This parameter is deprecated and will be removed.
         """
         raise TypeError("Distributed N-partite graph not supported")
